backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==1.10.12
pydantic-settings==2.1.0
elasticsearch==8.11.1
redis==5.0.1
faiss-cpu==1.7.4
langchain==0.1.0
python-dotenv==1.0.0
streamlit==1.30.0
requests==2.31.0